divided further into components
